Minor refactor for heroku cli validation script
